reorganized matching module
separated MatchingBase class from actual implementation of the Matchers
