+++ src/icepy4d/matching/__init__.py
@@ -1,1 +1,3 @@
-from .enums import Quality, GeometricVerification, TileSelection
+from .enums import Quality, GeometricVerification, TileSelection  # noqa: F401
+from .geometric_verification import geometric_verification  # noqa: F401
+from .matchers import *  # noqa: F401
