+++ src/icepy4d/matching/utils.py
@@ -1,108 +1,105 @@
-import importlib
 import logging
 import pickle
 from pathlib import Path
-from typing import Union
+from typing import List
 
 import cv2
 import numpy as np
 import torch
 
-import icepy4d.classes as icepy4d_classes
 from icepy4d.classes.features import Features
 
+# def geometric_verification(
+#     features: icepy4d_classes.FeaturesDict,
+#     threshold: float = 1,
+#     confidence: float = 0.9999,
+#     max_iters: int = 10000,
+#     laf_consistensy_coef: float = -1.0,
+#     error_type: str = "sampson",
+#     symmetric_error_check: bool = True,
+#     enable_degeneracy_check: bool = True,
+# ):
+#     """
+#     NOTE: this function is deprecated. Use geometric_verification
+#     from icepy4d.matching.geometric_verification instead.
 
-def geometric_verification(
-    features: icepy4d_classes.FeaturesDict,
-    threshold: float = 1,
-    confidence: float = 0.9999,
-    max_iters: int = 10000,
-    laf_consistensy_coef: float = -1.0,
-    error_type: str = "sampson",
-    symmetric_error_check: bool = True,
-    enable_degeneracy_check: bool = True,
-):
-    """
-    NOTE: this function is deprecated. Use geometric_verification
-    from icepy4d.matching.geometric_verification instead.
+#     Performs geometric verification of matches on full images using
+#     either Pydegensac or MAGSAC++.
 
-    Performs geometric verification of matches on full images using
-    either Pydegensac or MAGSAC++.
+#     Args:
+#         features (icepy4d_classes.FeaturesDict): A dictionary containing
+#         extracted features for each camera view.
+#             threshold (float): Pixel error threshold for considering a
+#             correspondence an inlier.
+#             confidence (float): The required confidence level in the results.
+#             max_iters (int): The maximum number of iterations for estimating
+#             the fundamental matrix.
+#             laf_consistensy_coef (float): The weight given to Local Affine
+#             Frame (LAF) consistency term for pydegensac.
+#             error_type (str): The error function used for computing the
+#             residuals in the RANSAC loop.
+#             symmetric_error_check (bool): If True, performs an additional check
+#             on the residuals in the opposite direction.
+#             enable_degeneracy_check (bool): If True, enables the check for
+#             degeneracy using SVD.
 
-    Args:
-        features (icepy4d_classes.FeaturesDict): A dictionary containing
-        extracted features for each camera view.
-            threshold (float): Pixel error threshold for considering a
-            correspondence an inlier.
-            confidence (float): The required confidence level in the results.
-            max_iters (int): The maximum number of iterations for estimating
-            the fundamental matrix.
-            laf_consistensy_coef (float): The weight given to Local Affine
-            Frame (LAF) consistency term for pydegensac.
-            error_type (str): The error function used for computing the
-            residuals in the RANSAC loop.
-            symmetric_error_check (bool): If True, performs an additional check
-            on the residuals in the opposite direction.
-            enable_degeneracy_check (bool): If True, enables the check for
-            degeneracy using SVD.
+#     Returns:
+#         None
 
-    Returns:
-        None
+#     Raises:
+#         ImportError: If Pydegensac library is not available.
 
-    Raises:
-        ImportError: If Pydegensac library is not available.
+#     Note:
+#         The function modifies directly the input Features object by removing
+#         the outliers after geoemtric verification.
 
-    Note:
-        The function modifies directly the input Features object by removing
-        the outliers after geoemtric verification.
+#     """
 
-    """
+#     logging.warning(
+#         "This function is deprecated. Use geometric_verification from icepy4. Use matching.geometric_verification instead."
+#     )
 
-    logging.warning(
-        "This function is deprecated. Use geometric_verification from icepy4. Use matching.geometric_verification instead."
-    )
+#     logging.info("Geometric verification of matches on full images...")
+#     cams = list(features.keys())
+#     try:
+#         pydegensac = importlib.import_module("pydegensac")
+#         use_pydegensac = True
+#     except ImportError:
+#         logging.error(
+#             "Pydegensac not available. Using MAGSAC++ (OpenCV) for geometric verification."
+#         )
+#         use_pydegensac = False
+#     if use_pydegensac:
+#         F, inlMask = pydegensac.findFundamentalMatrix(
+#             features[cams[0]].kpts_to_numpy(),
+#             features[cams[1]].kpts_to_numpy(),
+#             px_th=threshold,
+#             conf=confidence,
+#             max_iters=max_iters,
+#             laf_consistensy_coef=laf_consistensy_coef,
+#             error_type=error_type,
+#             symmetric_error_check=symmetric_error_check,
+#             enable_degeneracy_check=enable_degeneracy_check,
+#         )
+#         logging.info(
+#             f"Pydegensac found {inlMask.sum()} inliers ({inlMask.sum()*100/len(features[cams[0]]):.2f}%)"
+#         )
+#     else:
+#         F, inliers = cv2.findFundamentalMat(
+#             features[cams[0]].kpts_to_numpy(),
+#             features[cams[1]].kpts_to_numpy(),
+#             cv2.USAC_MAGSAC,
+#             0.5,
+#             0.999,
+#             100000,
+#         )
+#         inlMask = inliers > 0
+#         logging.info(
+#             f"MAGSAC++ found {inlMask.sum()} inliers ({inlMask.sum()*100/len(features[cams[0]].kpts_to_numpy()):.2f}%)"
+#         )
 
-    logging.info("Geometric verification of matches on full images...")
-    cams = list(features.keys())
-    try:
-        pydegensac = importlib.import_module("pydegensac")
-        use_pydegensac = True
-    except ImportError:
-        logging.error(
-            "Pydegensac not available. Using MAGSAC++ (OpenCV) for geometric verification."
-        )
-        use_pydegensac = False
-    if use_pydegensac:
-        F, inlMask = pydegensac.findFundamentalMatrix(
-            features[cams[0]].kpts_to_numpy(),
-            features[cams[1]].kpts_to_numpy(),
-            px_th=threshold,
-            conf=confidence,
-            max_iters=max_iters,
-            laf_consistensy_coef=laf_consistensy_coef,
-            error_type=error_type,
-            symmetric_error_check=symmetric_error_check,
-            enable_degeneracy_check=enable_degeneracy_check,
-        )
-        logging.info(
-            f"Pydegensac found {inlMask.sum()} inliers ({inlMask.sum()*100/len(features[cams[0]]):.2f}%)"
-        )
-    else:
-        F, inliers = cv2.findFundamentalMat(
-            features[cams[0]].kpts_to_numpy(),
-            features[cams[1]].kpts_to_numpy(),
-            cv2.USAC_MAGSAC,
-            0.5,
-            0.999,
-            100000,
-        )
-        inlMask = inliers > 0
-        logging.info(
-            f"MAGSAC++ found {inlMask.sum()} inliers ({inlMask.sum()*100/len(features[cams[0]].kpts_to_numpy()):.2f}%)"
-        )
-
-    features[cams[0]].filter_feature_by_mask(inlMask)
-    features[cams[1]].filter_feature_by_mask(inlMask)
+#     features[cams[0]].filter_feature_by_mask(inlMask)
+#     features[cams[1]].filter_feature_by_mask(inlMask)
 
 
 def process_resize(w, h, resize):
